refactor: organizing code and writing unit test
